safari-534.53-branch/LayoutTests/http/tests/websocket/tests/bufferedAmount-after-close-expected.txt

Web Socket bufferedAmount after closed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
Closed.
PASS ws.readyState is 2
PASS ws.bufferedAmount is 0
PASS ws.send('send to closed socket') is false
PASS ws.bufferedAmount is 23
PASS successfullyParsed is true

TEST COMPLETE
